falta

+++ practica_final3/parse.h
@@ -21,7 +21,7 @@ typedef struct
     int id_cliente;
     int codProducto;
     int cantidad;
-    int precioUnitario;
+    float precioUnitario;
 }eVentas;
